Modificada la aparición de disparos en el juego
El juego gestionará ahora una lista de disparos. Ahora mismo solo dispara una vez, un disparo y cada vez que se sale de la pantalla la lista de disparos sabe cuando tiene que eliminar ese disparo de la lista.

+++ core/src/com/tfg/spacegame/GameObject.java
@@ -10,7 +10,7 @@ public class GameObject {
     private Texture texture;
     private Rectangle logicShape;
 
-    public GameObject(String textureName, int x, int y) {
+    public GameObject(String textureName, float x, float y) {
         texture = AssetsManager.loadTexture(textureName);
 
         logicShape = new Rectangle();
